+++ src/components/behaviors/ShoutConfig.tsx
@@ -8,6 +8,7 @@ import { ShoutConfigModel } from '../../models/BotConfig'
 import ConfigTable from '../config/ConfigTable'
 import ConfigTableRow from '../config/ConfigTableRow'
 import { DefaultValuesChecker } from '../utils/DefaultValuesChecker'
+import TimeInput from '../config/TimeInput'
 
 
 type Props = {
@@ -34,7 +35,7 @@ const ShoutConfig = ({ config, onChange }: Props) => {
                     <ConfigTableRow
                         layout="v"
                         label={<ConfigLabel name="Interval" helpText="Interval between shouts in milliseconds." />}
-                        item={<NumericInput unit="ms" value={config.shout_interval} onChange={value => onChange?.({ ...config, shout_interval: value })} />}
+                        item={<TimeInput value={config.shout_interval} onChange={value => onChange?.({ ...config, shout_interval: value })} />}
                     />
                 </ConfigTable>
             </ConfigPanel>