+++ src/models/BotConfig.tsx
@@ -81,10 +81,8 @@ export type FarmingConfigModel = Partial<{
     is_stop_fighting: boolean;
     prevent_already_attacked: boolean;
 
-    obstacle_avoidance_enabled: boolean,
     obstacle_avoidance_cooldown: number,
     obstacle_avoidance_max_try: number,
-    obstacle_avoidance_only_passive: boolean,
 
     min_mobs_name_width: number,
     max_mobs_name_width: number,